Added create button to your cards page. Turned Create Card into Modal. Fixed difficulty state setting in QuizDifficulty.js so that state carries into quiz and clears on component mount
