+++ src/components/create/YourCards.js
@@ -4,6 +4,7 @@ import API from '../../modules/API/API'
 import UserSession from '../../modules/User/UserSession'
 import DeleteModal from './DeleteModal'
 import EditModal from './EditModal'
+import CreateModal from './CreateModal';
 
 export default class YourCards extends Component {
   state = {
@@ -30,6 +31,7 @@ render() {
       <Row>
         <h1>Your Cards</h1>
       </Row>
+      <CreateModal />
       <ListGroup className="d-flex inline">
         {
           this.state.userCards.map((card) =>
