+++ src/components/quiz/PeopleNeeded.js
@@ -8,7 +8,7 @@ render () {
     <Container>
       <Row>
         <h1>You need more cards to take this quiz.</h1>
-        <Button tag={Link} to="/create">Create Cards</Button>
+        <Button tag={Link} to="/cards">Create Cards</Button>
         <Button tag={Link} to="/quiz-select">Take Different Quiz</Button>
       </Row>
     </Container>
